Modify reflash to wait for reflash to finish before sending reset.
Requires that the new pic12_executive be loaded to work correctly.
If the old executive is loaded, it will throw an error saying
reflash not detected and quit.  The reflash is still working and
after ~20 seconds (<30 seconds) for the pic16lf1847, the reflash
will succeed.  This allows upgrading from an old pic12_executive
using the new modtool.