Fix streamers being added incorrectly to devices
pymomo had an incorrect format string for python struct module
that caused only the first character of the streamer name to be sent
in the rpc call rather than the entire thing.  This caused the name
not to match any module name and hence the stream would silently not
work.

also added ability to inspect the name of each streamer's destination
module using query_streamer