Reuse IOTileDeviceSlug from iotile_cloud
Main extra feature (apart from re-use) is that the device_id_to_slug
now supports portions of slugs as well: d—0123 or 5abc on top of the
integer.

May be handy as we begin to write Apps that interact with users that
may be think of the device ID as the one written on the box.

Ideally, we switch iotilecore overall to take either slugs or ids, so I
can also do `hw connect d—0123`