Add basic support for emulating physical devices.
This commit adds EmulatedDevice and EmulatedTile along with supporting
classes.  These two main classes inherit from VirtualIOTileDevice and
VirtualTile and are the base class for virtual devices that are designed
to emulate physical devices.

Emulated{Device,Tile} add support for tracking state changes and logging
to a file as well as dumping and loading their internal state and
easily loading prepared "scenarios" which will setup the internal
state of the emulated device/tile into a known configuration.