Add parallel rpc flow for combining the 2-byte feature+command in to a single command
rpc_v2 also enforces the use of arg_format and result_format, which changes the order for passing args through RPCs (python limitations mean we have to pass those arguments before *args and **kwargs, which are used in some rpc commands).

Right now it's only implemented for AdapterStream and CmdStream, but may need to add v2 methods to other streams if being used there as well.

Remove usage of RPCCommand