Refactor iotile-emulate to use asyncio event loop.
This is a major refactor that takes iotile-emulate from working
to working (which is why it is a single commit) but changes the
underlying emulation mechanism from relying on threading to
relying on asyncio.

Previously there needed to be a thread for each background task,
now there is a coroutine for each background task.  Because of
the pain of spinning up threads for everything, many subsystems
were wired together using a string of callbacks to perform
operations that needed to be asynchronous.  Now there is a single
EmulationLoop class that runs an asyncio event loop and any number
of coroutines can work cooperatively inside of the event loop.

Major Changes:
- Moved to asyncio from threading
- Dropped python 2 compatibility
- Refactored all controller subsystems to use a background coroutine
  for their background work instead of chained callbacks.
- Added proper support for background tasks on peripheral tiles.
- Added proper support for stopping background tasks on tiles when
  the emulated device is reset.
- Updated and adjusted module docstrings for accuracy given the
  new emulation model.
- Added support for implementing RPCs using coroutines and included
  a new @async_tile_rpc decorator.
- Remove all references to future and __future__ packages that are
  no longer needed since we don't target python 2.