Update UTCAssigner to be faster and more correct.
This commit changes UTCAssigner to memoize anchor points to
reduce the amount of work it needs to do to fix an entire report.

It also updates the utc assignment algorithm to look for reference
times in both directions (forward and backward in time) from a given
point that it wants to assign a timestamp to.

Closes #733