Refactor event lopo into async_tools and add OperationManager.
The OperationManager class is designed to simplify the job of
implementing complex, asynchronous message driven operations
by lessing coroutines block on the receipt of specific messages
with timeouts.

This blocking happens independently of the underlying message
transport layer so there is no "head of queue" blocking effect.

It will serve as the basis for many of the DeviceAdapter rewrites
as well as the Websocket client.