Major refactor of how DeviceAdapters work.
This commit lays the foundations for asynchronous device adapters.
It creates AbstractDeviceAdapter as the basic interface that all
adapters must implement and creates StandardDeviceAdapter as a
reference implementation of that interface.

There are two mixins that can be used to implement two of the
common things that DeviceAdapter classes are expected to support.

In order to provide backwards compatibility, there is a legacy
wrapper that can turn the new DeviceAdapter into the old one
so that it can be used unmodified by AdapterStream.