Refactor iotile-emulate to use BackgroundEventLoop
This is a minimal refactor to make EmulatedDevice work with
the SharedLoop architecture that is now present in iotile-core.

Previously EmulatedDevice spun up its own emulation loop in a different
thread with its own asyncio.Loop().  Now it just adds itself to the
chosen BackgroundEventLoop.

Note that there is still room for additional refactors to remove
some duplicate functionality in EmulationLoop but in the interest
of time, it is left for now.