Patch support for python 3.8.0 on windows.
There's a CPython bug that causes event loops running in background
threads to raise an exception on close.  The cause is that the
default event loop changed to ProactorEventLoop on windows and there
was a latent bug using a ProactorEventloop in a background thread.

This commit checks for this issue and patches the event loop policy
only on python 3.8.0 and windows to use the old SelectorEventloop.