Update Pylint config (#906)
* Move CLASSES section after IMPORTS (newer versions' style)

* Move CLASSES section after VARIABLES (newer versions' style)

* Update comments/style to newer pylint versions' approach

* Move TYPECHECK section after REFACTORING (newer versions' style)

* Add a new STRING section between REFACTORING and TYPECHECK

* Move SIMILARITIES section between REFACTORING and STRING

Newer versions' style.

* Move MISCELLANEOUS section after BASIC (newer versions' style)

* Update the disable line to a more readable format

Follows new pylint versions' style as well.

* Update default disabled warnings (new defaults in newer versions)

* Update redefining-builtins-modules with new default ones

* Replace old 'name-hint' settings with new 'naming-style' ones

* Drop rgx settings since they default to naming-style ones now

* Add BaseException to overgeneral ones (default in never versions)

* Add new sections from newer pylint versions

* Enable c-extension-no-member (newer pylint default)

* Add `logger` to good-names

For the module-level `logger` pattern used, for example, on archfx_cloud. E.g.

import logging

logger = logging.getLogger(__name__)

* Update deprecated-modules with new defaults

* Update valid-metaclass-classmethod-first-arg with a new default

* Add line length config for pycodestyle, pep8 and flake8

Usage of pycodestyle (former pep8) along with pylint is recommended in our style
guide. Flake8 (combines pycodestyle, pyflakes and some other minor scripts) is
what I personally use and a section for it should not cause any problems.

* Add classes with TestCase suffix to docstring check exclusion

Such classes are used extensively in our testing suites (for example, in
archfx_cloud) along with test_* methods.

* Require an LF line ending format

While not extremely important, and automatically followed 99% of the time, the
'same line ending everywhere' convention, when broken, can wreak havoc. It's
best to detect and prevent.

* Update deprecated-modules format for readability

* Add `imp` to deprecated-modules

It is, in fact, deprecated.

* Enable len-as-condition, redefined-builtin, redefined-outer-name

Feel like in all cases it would be better to use an explicit #disable comment to
further emphasize that the breach is intended.

* Revert "Add line length config for pycodestyle, pep8 and flake8"

This reverts commit c5b2a9378e7d85b5238092ef064715652fa394a7.