Add support for device servers to indicate fatal errors.
Currently, the virtual-device script has no way of knowing if the
underlying device server that it is hosting has encountered a
fatal error and is no longer running.  Since the server runs
in a background event loop, exceptions don't naturally cause the
main script to exit.

This commit adds a failed Future() on all DeviceServers that can
be set to indicate that the server has encounted a fatal error
condition and stopped operation.

The virtual-device script is updated to wait on this future rather
than just sleeping until Ctrl-C is hit.

An example of the new usage is:

    virtual-device bled112 simple

If you unplug the BLED112 dongle that is serving the virtual device,
it can now cleanly exit the virtual_device program.