Fix error handling in auth process
It assumed that functions communicate with BLE device return tuple
(bool, obj) where bool is True for function successfully communicated
with the device and obj contains data received or where bool is False
and obj is a dictionary with a reason.
Functions that handle auth handshake are updated to be compliant.