Fix commands hanging when bled112 dongle is removed.
Also add an explicit `heartbeat` command to HardwareManager that can
be called to check if the underlying DeviceAdapter is still functional.

This is a bit of a hack but doesn't fundamentally break the architecture
and is a placeholder until we finish porting the bled112 device adapter
over to a modern AbstractDeviceAdapter where this functionality would
come for free.